+++ build.bat
@@ -42,9 +42,12 @@ if %ERRORLEVEL% NEQ 0 exit /b 1
 gcc -Wall -Wextra -std=c99 -g -O0 -D_GNU_SOURCE -c src\semantic\register_allocator.c -o obj\semantic\register_allocator.o
 if %ERRORLEVEL% NEQ 0 exit /b 1
 
-echo Compiling code generator...
-gcc -Wall -Wextra -std=c99 -g -O0 -D_GNU_SOURCE -c src\codegen\code_generator.c -o obj\codegen\code_generator.o
-if %ERRORLEVEL% NEQ 0 exit /b 1
+echo Compiling code generator modules...
+for %%f in (src\\codegen\\*.c) do (
+    echo   %%~nxf
+    gcc -Wall -Wextra -std=c99 -g -O0 -D_GNU_SOURCE -c %%f -o obj\\codegen\\%%~nf.o
+    if errorlevel 1 exit /b 1
+)
 
 echo Compiling debug info...
 gcc -Wall -Wextra -std=c99 -g -O0 -D_GNU_SOURCE -c src\debug\debug_info.c -o obj\debug\debug_info.o
@@ -63,7 +66,7 @@ gcc -Wall -Wextra -std=c99 -g -O0 -D_GNU_SOURCE -c src\main.c -o obj\main.o
 if %ERRORLEVEL% NEQ 0 exit /b 1
 
 echo Linking...
-gcc obj\lexer\lexer.o obj\parser\ast.o obj\parser\parser.o obj\semantic\symbol_table.o obj\semantic\type_checker.o obj\semantic\register_allocator.o obj\codegen\code_generator.o obj\debug\debug_info.o obj\runtime\gc.o obj\error\error_reporter.o obj\main.o -o bin\methasm.exe
+gcc obj\lexer\lexer.o obj\parser\ast.o obj\parser\parser.o obj\semantic\symbol_table.o obj\semantic\type_checker.o obj\semantic\register_allocator.o obj\\codegen\\*.o obj\debug\debug_info.o obj\runtime\gc.o obj\error\error_reporter.o obj\main.o -o bin\methasm.exe
 
 if %ERRORLEVEL% EQU 0 (
     echo Build successful! Executable created at bin\methasm.exe
@@ -71,3 +74,5 @@ if %ERRORLEVEL% EQU 0 (
     echo Build failed!
     exit /b 1
 )
+
+
